+++ libs/@guardian/atoms-rendering/src/expandableAtom/Container.tsx
@@ -1,5 +1,7 @@
 import { css } from '@emotion/react';
+import type { SerializedStyles } from '@emotion/react';
 import type { ArticleTheme } from '@guardian/libs';
+import { ArticleDesign } from '@guardian/libs';
 import { neutral, text } from '@guardian/source-foundations';
 import { Summary } from './Summary';
 
@@ -8,40 +10,46 @@ const containerStyling = css`
 	position: relative;
 `;
 
-const detailStyling = css`
-	margin: 16px 0 36px;
-	background: ${neutral[93]};
-	color: ${text.primary};
-	padding: 0 5px 6px;
-	border-image: repeating-linear-gradient(
-			to bottom,
-			${neutral[86]},
-			${neutral[86]} 1px,
-			transparent 1px,
-			transparent 4px
-		)
-		13;
-	border-top: 13px solid black;
-	position: relative;
-	summary {
-		list-style: none;
-		margin: 0 0 16px;
-	}
+export const detailStyling = (design?: ArticleDesign): SerializedStyles => {
+	// One off background colour for analysis articles
+	const background =
+		design === ArticleDesign.Analysis ? '#F2E8E6' : neutral[93];
+	return css`
+		margin: 16px 0 36px;
+		background: ${background};
+		color: ${text.primary};
+		padding: 0 5px 6px;
+		border-image: repeating-linear-gradient(
+				to bottom,
+				${neutral[86]},
+				${neutral[86]} 1px,
+				transparent 1px,
+				transparent 4px
+			)
+			13;
+		border-top: 13px solid black;
+		position: relative;
+		summary {
+			list-style: none;
+			margin: 0 0 16px;
+		}
 
-	/* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#Customizing_the_disclosure_widget */
-	summary::-webkit-details-marker {
-		display: none;
-	}
+		/* https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#Customizing_the_disclosure_widget */
+		summary::-webkit-details-marker {
+			display: none;
+		}
 
-	summary:focus {
-		outline: none;
-	}
-`;
+		summary:focus {
+			outline: none;
+		}
+	`;
+};
 
 export const Container = ({
 	id,
 	title,
 	children,
+	design,
 	pillar,
 	expandForStorybook,
 	atomType,
@@ -50,6 +58,7 @@ export const Container = ({
 }: {
 	id: string;
 	title: string;
+	design?: ArticleDesign;
 	pillar: ArticleTheme;
 	expandForStorybook?: boolean;
 	atomType: string;
@@ -59,7 +68,7 @@ export const Container = ({
 }): JSX.Element => (
 	<div css={containerStyling} data-atom-id={id} data-atom-type={atomType}>
 		<details
-			css={detailStyling}
+			css={detailStyling(design)}
 			data-atom-id={id}
 			data-snippet-type={atomType}
 			open={expandForStorybook}