+++ libs/@guardian/eslint-plugin-source-foundations/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @guardian/eslint-plugin-source-foundations
 
+## 9.0.3
+
+### Patch Changes
+
+- 7cd6f21: Bump tslib to v2.4.1
+
+## 9.0.2
+
+### Patch Changes
+
+- cf437f9: Bump @guardian/source-foundations to v8
+
+## 9.0.1
+
+### Patch Changes
+
+- d742b50: Bump @guardian/libs to 12.0.0
+
 ## 9.0.0
 
 ### Major Changes