+++ libs/@guardian/source-react-components-development-kitchen/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @guardian/source-react-components-development-kitchen
 
+## 8.2.5
+
+### Patch Changes
+
+- 7cd6f21: Bump tslib to v2.4.1
+
+## 8.2.4
+
+### Patch Changes
+
+- cf437f9: Bump @guardian/source-foundations to v8
+- Updated dependencies [cf437f9]
+  - @guardian/source-react-components@10.0.1
+
+## 8.2.3
+
+### Patch Changes
+
+- d742b50: Bump @guardian/libs to 12.0.0
+
+## 8.2.2
+
+### Patch Changes
+
+- 0fcf574: ExpandingWrapper - update styling and allow cssOverrides
+
+## 8.2.1
+
+### Patch Changes
+
+- c882f90: Update styling for expanding wrapper
+- 6ae535a: Update fileinput styling and add hover styling
+
 ## 8.2.0
 
 ### Minor Changes